+++ apps/frontend/src/routes/_auth/administration/export.tsx
@@ -1,5 +1,17 @@
-import { createFileRoute } from "@tanstack/react-router";
+import { Paper, Title } from "@quassel/ui";
+import { createFileRoute, Outlet } from "@tanstack/react-router";
+
+function AdministrationExport() {
+  return (
+    <>
+      <Title>Export</Title>
+      <Paper my="lg">
+        <Outlet />
+      </Paper>
+    </>
+  );
+}
 
 export const Route = createFileRoute("/_auth/administration/export")({
-  component: () => <div>Hello /_auth/administration/export!</div>,
+  component: AdministrationExport,
 });