+++ apps/frontend/src/routes/_auth/questionnaire/_questionnaire/new.tsx
@@ -4,6 +4,7 @@ import { useStore } from "@nanostores/react";
 import { PeriodForm, PeriodFormValues } from "../../../../components/questionnaire/PeriodForm";
 import { $api } from "../../../../stores/api";
 import { $questionnaire } from "../../../../stores/questionnaire";
+import { useEffect } from "react";
 
 const messages = i18n("questionnaireNew", {
   title: "Create new period of life",
@@ -16,11 +17,21 @@ function QuestionnaireNew() {
 
   const questionnaire = useStore($questionnaire);
 
+  const { data: participant } = $api.useQuery("get", "/participants/{id}", {
+    params: { path: { id: questionnaire!.participant.id.toString() } },
+  });
+
+  useEffect(() => {
+    if (participant) $questionnaire.set({ ...questionnaire!, participant });
+  }, [participant]);
+
   const createQuestionnaireMutation = $api.useMutation("post", "/questionnaires", {
     onSuccess: (questionnaire) => {
       n({ to: "/questionnaire/$id/entries", params: { id: questionnaire.id.toString() } });
     },
   });
+
+  const prevEndDate = questionnaire?.participant.latestQuestionnaire?.endedAt;
 
   const onSave = (form: PeriodFormValues) => {
     const {
@@ -28,8 +39,8 @@ function QuestionnaireNew() {
       range: [localStartedAt, localEndedAt],
     } = form;
 
-    const startedAt = localStartedAt.toISOString();
-    const endedAt = localEndedAt.toISOString();
+    const startedAt = localStartedAt!.toISOString();
+    const endedAt = localEndedAt!.toISOString();
 
     createQuestionnaireMutation.mutate({
       body: { title, startedAt, endedAt, study: questionnaire!.study.id, participant: questionnaire!.participant.id },
@@ -39,7 +50,7 @@ function QuestionnaireNew() {
   return (
     <>
       <h3>{t.title}</h3>
-      <PeriodForm onSave={onSave} actionLabel={t.formAction} />
+      <PeriodForm onSave={onSave} prevEndDate={prevEndDate ? new Date(prevEndDate) : undefined} actionLabel={t.formAction} />
     </>
   );
 }