+++ libs/ui/src/components/MonthPicker.tsx
@@ -19,7 +19,7 @@ export function MonthPicker<Type extends DatePickerType = "default">({ selectEnd
   const onChange: MonthPickerProps<Type>["onChange"] = (date) => {
     if (rest.type === "range" && isRangeValue(date) && selectEndOfMonth) {
       const [start, end] = date;
-      const newDate = [start, dayjs(end).utc().endOf("month").toDate()];
+      const newDate = [start, end ? dayjs(end).utc().endOf("month").toDate() : undefined];
       date = newDate as typeof date;
     }
     rest.onChange!(date);